Add module for extracting VOWL css to d3 rules
When using the export function, all VOWL css rules have to be inlined
and later be removed again. The collection of this rules takes too long
if you do it manually.
This tool collects them and prints them to the command line. They can be
copied and have to be placed manually in the correct position in the
according webvowl-app file.