Use smarter equivalent property range merging
Closes #28
When merging ranges of equivalent properties, owl:Thing was always chosen
for Object properties. From now on it will be looked whether, exactly 1
range exists which is different from owl:Thing (that means a range which
isn't that general). If thats the case, this range will be used as merge
range, otherwise owl:Thing.